requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
python-multipart==0.0.6
google-genai==1.28.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.1.1
pyshp==2.3.1 